requirements.txt
Flask==2.2.3
sentencepiece==0.1.98
transformers==4.33.2
gunicorn==20.1.0
Flask-Cors==3.0.10
scipy==1.10.1
timm==0.9.7
Werkzeug==2.2.2

# OD model -------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation